refactor(core): split soft_failure into three focused modules
Decompose the monolithic `core/soft_failure.py` module into three
single-responsibility modules so each consumer imports only the
concept it actually uses:

- `core/errors.py` — houses `NoStructuredOutputError`, the peer
  exception of `AgentSDKError` shared by chains, adapters, and the
  retry policy.
- `core/stream_drain.py` — houses the `drain` two-call helper that
  collapses the duplicated `ResultEvent` collection pattern at every
  soft-failure raise site.
- `core/error_egress.py` — houses `_redact_credentials` and
  `build_error_event`, the egress-side mapping from exceptions to the
  wire-visible `ErrorEvent` (with credential scrubbing).

This sharpens module boundaries: adapters that only need credential
redaction no longer transitively import the drain helper, and chains
that only need `drain` + `NoStructuredOutputError` no longer pull in
the egress mapper. All call sites and tests are updated to the new
import paths, the obsolete `tests/core/test_soft_failure.py` is
replaced by `tests/core/test_error_egress.py`, and the `RaiseSite`
doc references in `types/domain/agent.py` are repointed at
`core/errors.py`. No behavioural change.
