+++ src/kodezart/adapters/claude_agent_executor.py
@@ -12,8 +12,8 @@ from claude_agent_sdk import (
 
 from kodezart.adapters._permission_modes import _validate_permission_mode
 from kodezart.adapters._sdk_mapping import map_message
+from kodezart.core.error_egress import _redact_credentials
 from kodezart.core.logging import BoundLogger, get_logger
-from kodezart.core.soft_failure import _redact_credentials
 from kodezart.domain.errors import AgentSDKError
 from kodezart.types.domain.agent import AgentEvent
